backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
mysqlclient>=2.2
django-cors-headers>=4.0
gunicorn>=20.1
Pillow>=9.5
